extract synth_one_iteration
